Expose the check function
