+++ pkg/codegen/go/gen_program_test.go
@@ -2,9 +2,8 @@ package gen
 
 import (
 	"bytes"
-	"fmt"
 	"io/ioutil"
-	"os"
+
 	"path/filepath"
 	"testing"
 
@@ -17,8 +16,6 @@ import (
 	"github.com/pulumi/pulumi/pkg/v3/codegen/pcl"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/testing/test"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/testing/utils"
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/executable"
 )
 
 var testdataPath = filepath.Join("..", "testing", "test", "testdata")
@@ -29,7 +26,9 @@ func TestGenerateProgram(t *testing.T) {
 			Language:   "go",
 			Extension:  "go",
 			OutputFile: "main.go",
-			Check:      goCheck,
+			Check: func(t *testing.T, path string, dependencies codegen.StringSet) {
+				Check(t, path, dependencies, "../../../../../../../sdk")
+			},
 			GenProgram: GenerateProgram,
 			TestCases:  test.PulumiPulumiProgramTests,
 		})
@@ -82,41 +81,4 @@ func newTestGenerator(t *testing.T, testFile string) *generator {
 	}
 	g.Formatter = format.NewFormatter(g)
 	return g
-}
-
-func goCheck(t *testing.T, path string, _ codegen.StringSet) {
-	dir := filepath.Dir(path)
-	ex, err := executable.FindExecutable("go")
-	require.NoError(t, err)
-
-	// We remove go.mod to ensure tests are reproducible.
-	goMod := filepath.Join(dir, "go.mod")
-	if err = os.Remove(goMod); !os.IsNotExist(err) {
-		require.NoError(t, err)
-	}
-	err = integration.RunCommand(t, "generate go.mod",
-		[]string{ex, "mod", "init", "main"},
-		dir, &integration.ProgramTestOptions{})
-	require.NoError(t, err)
-	err = integration.RunCommand(t, "go tidy",
-		[]string{ex, "mod", "tidy"},
-		dir, &integration.ProgramTestOptions{})
-	require.NoError(t, err)
-	err = integration.RunCommand(t, "point towards local Go SDK",
-		[]string{ex, "mod", "edit",
-			fmt.Sprintf("--replace=%s=%s",
-				"github.com/pulumi/pulumi/sdk/v3",
-				"../../../../../../../sdk")},
-		dir, &integration.ProgramTestOptions{})
-	require.NoError(t, err)
-	err = integration.RunCommand(t, "go tidy after replace",
-		[]string{ex, "mod", "tidy"},
-		dir, &integration.ProgramTestOptions{})
-	require.NoError(t, err)
-
-	err = integration.RunCommand(t, "test build", []string{ex, "build", "-v", "all"},
-		dir, &integration.ProgramTestOptions{})
-	require.NoError(t, err)
-	os.Remove(filepath.Join(dir, "main"))
-	assert.NoError(t, err)
 }
