+++ pkg/codegen/python/gen_program_test.go
@@ -4,12 +4,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
-	"github.com/pulumi/pulumi/pkg/v3/codegen"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/testing/test"
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-	"github.com/pulumi/pulumi/sdk/v3/python"
 )
 
 var testdataPath = filepath.Join("..", "testing", "test", "testdata")
@@ -20,18 +15,8 @@ func TestGenerateProgram(t *testing.T) {
 			Language:   "python",
 			Extension:  "py",
 			OutputFile: "__main__.py",
-			Check:      pythonCheck,
+			Check:      Check,
 			GenProgram: GenerateProgram,
 			TestCases:  test.PulumiPulumiProgramTests,
 		})
-}
-
-func pythonCheck(t *testing.T, path string, _ codegen.StringSet) {
-	ex, _, err := python.CommandPath()
-	require.NoError(t, err)
-	name := filepath.Base(path)
-	dir := filepath.Dir(path)
-	err = integration.RunCommand(t, "python syntax check",
-		[]string{ex, "-m", "py_compile", name}, dir, &integration.ProgramTestOptions{})
-	require.NoError(t, err)
 }
